refactor: cleaning auth.ts, moving everything to auth.config.ts so can be used in middleware.ts
